+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 100.2.1
+
+- Fix review form persistence
+- Simpler way to add an action to validate
+- New data persistor interface
+
 ## 100.2.0
 
 - Admin "login" and "reset password" forms validation